+++ Autumn22/004.Stack/tests/testMain.cpp
@@ -43,7 +43,8 @@ int main(int argc, const char* const* const argv) {
     stack::push(&stack, 5);
     //STACK__dump(stack));
     //int var = 0;
-    uint8_t var = 0;
+    //uint8_t var = 0;
+    uint32_t var = 0;
     stack::pop(&stack, &var);
     //STACK__dump(stack));
     stack::pop(&stack);
